requirements.txt
BeautifulSoup==3.2.1
Django==1.3.1
Reindent==0.1.1
Scrapy==0.14.1
South==0.7.3
Twisted==12.0.0
amqplib==1.0.2
anyjson==0.3.1
celery==2.5.5
celery-with-redis==2.5
django-celery==2.5.5
django-model-utils==1.1.0
django-picklefield==0.2.1
django-polymorphic==0.2
django-tastypie==0.9.11
html2text==3.200.3
kombu==2.1.8
lxml==2.3.4
mimeparse==0.1.3
python-dateutil==1.5
redis==2.4.13
simplejson==2.2.1
w3lib==1.2
wsgiref==0.1.2
zope.interface==4.0.1
